ME-639: Moved MockDocument into test_room.py.
